delete problem solve

+++ admin/ajax/features_facilities.php
@@ -90,10 +90,20 @@ if (isset($_POST['get_facilities'])) {
 
 if (isset($_POST['rem_facility'])) {
     $frm_data = filteration($_POST);
-    $q = "DELETE FROM `facilities` WHERE `id`=?";
     $values = [$frm_data['rem_facility']];
-    $res = delete($q, $values, 'i');
-    echo $res;
+    
+
+    $pre_q = "SELECT * FROM `facilities` WHERE `id`=?";
+    $res = select($pre_q, $values, 'i');
+    $img = mysqli_fetch_assoc($res);
+     
+    if (deleteImage($img['icon'], FACILITIES_FLODER)) {
+        $q = "DELETE FROM `facilities` WHERE `id`=?";
+        $res = delete($q, $values, 'i');
+        echo $res;
+    } else {
+        echo 0;
+    }
 }
